Divide portal into two toolbars.
